// apps/web/src/interface/user.interface.ts
export interface IUsers {
  userEmail: string;
  userPassword: string;
  roleId?: number;
  userName?: string;
}

// export interface IUsers {
//   userEmail: string;
//   userPassword: string;
//   userName: string;
//   userReferralCode: string;
//   roleId: number;
//   isVerified: boolean;
// }